# Make get_or_set_index immune to memory address reuse